refactor(auth): extract MFA/AAL redirect orchestration into useAuthRedirect hook
Move the race-prone post-auth redirect logic (MFA challenge gate, OAuth
mfa_required handoff, AAL2 elevation double-check, passkey fast-path,
?next= return URL, session-expired message) out of page.tsx into a
single documented hook. Behavior-identical; page.tsx 163->32 lines.
No visual change.
